requirements.txt
torch
torchvision
hydra-core
monai[all]
accelerate
tqdm
wandb
pycad-medic